requirements.txt
django==4.2.0
djangorestframework
django-cors-headers
mysqlclient
djangorestframework-simplejwt
djangochannelsrestframework